Add labelmap tools as well
